chore: Update subproject commit in xira
This commit updates the subproject commit in the xira file. The previous commit was "c10359644544baaf174b8b572dd6c22fe0738ba4" and it has been modified to "c10359644544baaf174b8b572dd6c22fe0738ba4-dirty". This change ensures that the correct subproject commit is used, improving the accuracy of the codebase.